refactor: extract option button
