+++ components/OptionButton/OptionButton.stories.tsx
@@ -1,0 +1,22 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { noop } from 'lodash'
+import { Skull } from 'lucide-react-native'
+import OptionButton from './OptionButton'
+
+const OptionButtonMeta: Meta<typeof OptionButton> = {
+  title: 'OptionButton',
+  component: OptionButton,
+  args: {
+    disabled: false,
+    icon: Skull,
+    loading: false,
+    onPress: noop,
+    title: 'Option button'
+  }
+}
+
+export default OptionButtonMeta
+
+type Story = StoryObj<typeof OptionButton>
+
+export const Default: Story = {}
