+++ components/OptionButton/OptionButton.tsx
@@ -1,18 +1,24 @@
-import { Button, Text, VStack } from 'appdeptus/components'
 import { type LucideIcon } from 'lucide-react-native'
+import Button from '../Button'
+import Text from '../Text'
+import { VStack } from '../ui'
 
 type OptionButtonProps = {
   icon: LucideIcon
   onPress: () => void
   title: string
+
+  disabled?: boolean
+  loading?: boolean
 }
 
-const OptionButton = ({ icon, onPress, title }: OptionButtonProps) => (
+const OptionButton = ({ icon, onPress, title, ...rest }: OptionButtonProps) => (
   <VStack
     className='items-center'
     space='xs'
   >
     <Button
+      {...rest}
       onPress={onPress}
       variant='callback'
       icon={icon}
