+++ components/index.ts
@@ -11,6 +11,7 @@ export { default as Card } from './Card'
 export { default as CardMenu } from './CardMenu'
 export { default as Disclaimer } from './Disclaimer'
 export { default as Dots } from './Dots'
+export { default as DoubleDataTable } from './DoubleDataTable'
 export { default as EmptyListItem } from './EmptyListItem'
 export { default as EnhancementListItem } from './EnhancementListItem'
 export { default as Error } from './Error'
@@ -18,18 +19,19 @@ export { default as ErrorBoundary } from './ErrorBoundary'
 export { default as FilterTopBar } from './FilterTopBar'
 export { default as GameArmyRoster } from './GameArmyRoster'
 export { default as GameBackground } from './GameBackground'
-export { default as GameDataTable } from './GameDataTable'
 export { default as IconBadge } from './IconBadge'
 export { default as Input } from './Input'
 export { default as InsetShadow } from './InsetShadow'
 export { default as Loading } from './Loading'
 export { default as NavigationHeader } from './NavigationHeader'
+export { default as OptionButton } from './OptionButton'
 export { default as PlayerTag } from './PlayerTag'
 export { default as Progress } from './Progress'
 export { default as Scoreboard } from './Scoreboard'
 export { default as ScreenContainer } from './ScreenContainer'
 export { default as ScreenSubtitle } from './ScreenSubtitle'
 export { default as ScreenTitle } from './ScreenTitle'
+export { default as SingleDataTable } from './SingleDataTable'
 export { default as TabMenu } from './TabMenu'
 export { default as Text } from './Text'
 export { default as TextLink } from './TextLink'
